Revert "corrected routes of menteeDashboard"
