+++ src/components/Mentee/Sidebar.js
@@ -1,82 +1,76 @@
-//import useState hook to create menu collapse state
-import React, { useState } from "react";
-
-
-//import react pro sidebar components
-import {
-  ProSidebar,
-  Menu,
-  MenuItem,
-  SidebarHeader,
-  SidebarFooter,
-  SidebarContent,
-} from "react-pro-sidebar";
-
-//import icons from react icons
-import { FaList, FaRegHeart } from "react-icons/fa";
-import {
-  FiHome,
-  FiLogOut,
-  FiArrowLeftCircle,
-  FiArrowRightCircle,
-} from "react-icons/fi";
-import { RiPencilLine } from "react-icons/ri";
-import { BiCog } from "react-icons/bi";
-
-//import sidebar css from react-pro-sidebar module and our custom css
-import "react-pro-sidebar/dist/css/styles.css";
-import "./MenteeDashboard.scss";
-
+import React, { useState } from 'react';
+import MenteeDashboard from './MenteeDashboard';
+import Select from 'react-select'
+import { Browse } from './Browse';
+import "./Browse.scss"
 const Sidebar = () => {
-  
-  
-  //create initial menuCollapse state using useState hook
-  const [menuCollapse, setMenuCollapse] = useState(false);
-
-  //create a custom function that will change menucollapse state from false to true and true to false
-  const menuIconClick = () => {
-    //condition checking to change state from true to false and vice versa
-    menuCollapse ? setMenuCollapse(false) : setMenuCollapse(true);
-  };
-
+  const options = [
+    { value: 'IAS', label: 'IAS' },
+    { value: 'JEE', label: 'JEE' },
+    { value: 'NEET', label: 'NEET' },
+    { value: 'Engineer', label: 'Engineer'},
+    { value: 'M.tech', label: 'M.tech'}
+  ]
+  const sty = {
+    width: "25em",
+    marginLeft: "25em"
+  }
+  const [results, setresults] = useState(<div></div>);
+  const onsele = async (e) => {
+    const que  = e.value;
+    const response = await fetch("http://localhost:5000/api/search", {
+      method: "POST",
+      headers: {
+        "Content-Type": "application/json",
+      },
+      body: JSON.stringify({
+        query: que,
+      }),
+    });
+    const res = await response.json();
+    console.log(res.finds);
+    const f = res.finds;
+    const r = f.map( (find) => 
+    <div className="Browse" style={{margin: "1em 1.2em 1em 1em"}}>
+    <div class="card-container">
+	{/* <span class="pro">PRO</span> */}
+	<img class="round" src="https://randomuser.me/api/portraits/women/79.jpg" alt="user" />
+	<h3>{find.name}</h3>
+	<h6>New York</h6>
+	{/* <p>User interface designer and <br/> front-end developer</p> */}
+	<div class="buttons">
+		<button class="primary">
+			Book A Session
+		</button>
+		{/* <button class="primary ghost">
+			Following
+		</button> */}
+	</div>
+	<div class="skills">
+		<h6>Mentorship In</h6>
+		<ul>
+			<li>{find.qualification}</li>
+			{/* <li>HTML</li>
+			<li>CSS</li>
+			<li>JavaScript</li>
+			<li>React</li>
+			<li>Node</li> */}
+		</ul>
+	</div>
+</div>       
+    </div>
+    );
+    console.log(r);
+    setresults(r);
+  }
   return (
-    <>
-      <div className="MenteeDashboard">
-      <div id="header">
-        {/* collapsed props to change menu size using menucollapse state */}
+    <div>
+        <MenteeDashboard />
+        <div style={sty}><Select options={options} onChange={onsele}/></div>
+        <div style={{margin: "1em 0em 0em 10em", display: "flex", flexWrap: "wrap"}}>{results}</div>
         
-        <ProSidebar collapsed={menuCollapse}>
-          <SidebarHeader>
-            <div className="logotext">
-              {/* small and big change using menucollapse state */}
-              <p>{menuCollapse ? "Logo" : "Big Logo"}</p>
-            </div>
-            <div className="closemenu" onClick={menuIconClick}>
-              {/* changing menu collapse icon on click */}
-              {menuCollapse ? <FiArrowRightCircle /> : <FiArrowLeftCircle />}
-            </div>
-          </SidebarHeader>
-          <SidebarContent>
-            <Menu iconShape="square">
-              <MenuItem active={true} icon={<FiHome />}>
-                Home
-              </MenuItem>
-              <MenuItem icon={<FaList />}>Messages</MenuItem>
-              <MenuItem icon={<FaRegHeart />}>Browse</MenuItem>
-              <MenuItem icon={<RiPencilLine />}>Profile</MenuItem>
-              <MenuItem icon={<BiCog />}>Settings</MenuItem>
-            </Menu>
-          </SidebarContent>
-          <SidebarFooter>
-            <Menu iconShape="square">
-              <MenuItem icon={<FiLogOut />}>Logout</MenuItem>
-            </Menu>
-          </SidebarFooter>
-        </ProSidebar>
-      </div>
-      </div>
-    </>
-  );
+    </div>
+  )
 };
 
 export default Sidebar;
